wip: case-sensitive-find.js: continue to disect & clean up the code; replace decrementing for loops (`i--;`) w/incrementing ones
Note: this may be a bug I'm introducing by using an incrementing for loop instead of a
decrementing one. I'm not sure yet.

Also, I've commented out some stuff in the `clearEverything()` func,
which may also be a bug (the act of commenting out I mean) and need to
be uncommented again in the future! Not sure yet yet.